requirements.txt
streamlit
ultralytics
pillow
opencv-python
onnx
onnxruntime
numpy
pandas
